[CHORE] Refactor tests to have a BaseModel they inherit from
