+++ tests/test_question_answers.py
@@ -1,68 +1,9 @@
 '''tests/test_question_answers.py'''
-import os
-import unittest
 import json
-import psycopg2
-from app.application import create_app
-from instance.config import app_config
+from tests.base_test import BaseTest
 
-class TestQuestionAnswers(unittest.TestCase):
+class TestQuestionAnswers(BaseTest):
     '''Class Testing Question Answers'''
-    def setUp(self):
-        self.app = create_app(config_name="testing")
-        self.client = self.app.test_client
-
-        #signup 3 users
-        self.client().post('/api/v1/auth/signup',
-                           content_type="application/json",
-                           data=json.dumps({"name": "Njeri", "username": "njery",
-                                            "email": "njeri@to.com", "password": "Test123",
-                                            "confirm_password": "Test123"}))
-        self.client().post('/api/v1/auth/signup',
-                           content_type="application/json",
-                           data=json.dumps({"name": "Suite", "username": "suite",
-                                            "email": "suite@to.com", "password": "Test123",
-                                            "confirm_password": "Test123"}))
-        self.client().post('/api/v1/auth/signup',
-                           content_type="application/json",
-                           data=json.dumps({"name": "Voter", "username": "voter",
-                                            "email": "voter@to.com", "password": "Test123",
-                                            "confirm_password": "Test123"}))
-        #login in all users
-        res = self.client().post('/api/v1/auth/login', content_type="application/json",
-                                 data=json.dumps({"username": "njery", "password": "Test123"}))
-        u_data = json.loads(res.data)
-        self.a_token = u_data["token"]
-
-        res2 = self.client().post('/api/v1/auth/login', content_type="application/json",
-                                  data=json.dumps({"username": "suite", "password": "Test123"}))
-        u_data2 = json.loads(res2.data)
-        self.a_token2 = u_data2["token"]
-        #user1 posts 3 questions
-        self.client().post('/api/v1/questions',
-                           headers=dict(Authorization="Bearer " + self.a_token),
-                           content_type="application/json",
-                           data=json.dumps({"title": "Git branching",
-                                            "content": "How to create & checkout a branch in git"}))
-        self.client().post('/api/v1/questions',
-                           headers=dict(Authorization="Bearer " + self.a_token),
-                           content_type="application/json",
-                           data=json.dumps({"title": "Baking",
-                                            "content": "How much baking soda to use?"}))
-        self.client().post('/api/v1/questions',
-                           headers=dict(Authorization="Bearer " + self.a_token),
-                           content_type="application/json",
-                           data=json.dumps({"title": "Sample",
-                                            "content": "Content"}))
-        #user2 posts answers
-        self.client().post('/api/v1/questions/1/answers',
-                           headers=dict(Authorization="Bearer " + self.a_token2),
-                           content_type="application/json",
-                           data=json.dumps({"content": "Use git branch <branch_name>"}))
-        self.client().post('/api/v1/questions/1/answers',
-                           headers=dict(Authorization="Bearer " + self.a_token2),
-                           content_type="application/json",
-                           data=json.dumps({"content": "Sample answer"}))
     def test_post_answer(self):
         '''test handling posting an answer'''
         #test successful post
@@ -137,13 +78,4 @@ class TestQuestionAnswers(unittest.TestCase):
         my_data4 = json.loads(result4.data)
         self.assertEqual(len(my_data4), 1)
         self.assertEqual(result4.status_code, 200)
-
-    def tearDown(self):
-        current_environemt = os.environ['ENV']
-        conn_string = app_config[current_environemt].CONNECTION_STRING
-        conn = psycopg2.connect(conn_string)
-        cursor = conn.cursor()
-        cursor.execute("DROP TABLE votes, comments, answers, questions, revoked_tokens, users")
-        conn.commit()
-        conn.close()
